@rrrlw 2021-04-28
P.S. - I don't think `calc_np_all2` was in `NAMESPACE`, might have to run `devtools::document` to test